Resolves a conflict in MultiIndex._level_index_from_level, which main
rewrote in #23370 to fix the same negative-level wraparound. Keeps
main's `norm` formulation and layers on the two things pandas'
MultiIndex._get_level_number does that neither side had: distinct
wording for underflow ("-3 is not a valid level number") versus
overflow ("not 3"), and the ValueError for an ambiguous duplicate
level name.

Also restores _check_duplicate_level_names, which this branch had
removed as unused. Main added two callers of it in the meantime
(DataFrame.stack and reshape's unstack), and the textual merge took
the deletion without conflicting.

+++ cpp/src/io/orc/writer_impl.hpp
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: Copyright (c) 2019-2026, NVIDIA CORPORATION.
+ * SPDX-FileCopyrightText: Copyright (c) 2019-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
 
@@ -23,6 +23,7 @@
 #include <cuda/iterator>
 #include <thrust/host_vector.h>
 
+#include <cstdint>
 #include <memory>
 #include <string>
 #include <vector>
@@ -169,7 +170,7 @@ struct persisted_statistics {
     num_rows = 0;
   }
 
-  void persist(int num_table_rows,
+  void persist(uint64_t num_table_rows,
                single_write_mode write_mode,
                intermediate_statistics&& intermediate_stats,
                rmm::cuda_stream_view stream);
@@ -179,7 +180,7 @@ struct persisted_statistics {
   std::vector<rmm::device_uvector<char>> string_pools;
   std::vector<statistics_dtype> stats_dtypes;
   std::vector<data_type> col_types;
-  int num_rows = 0;
+  uint64_t num_rows = 0;
 };
 
 /**
